test database

+++ admin/database_utils.py
@@ -21,12 +21,14 @@ def create_table_discogs(connection):
     """Create the discogs table."""
     execute_sql(connection, "CREATE TABLE IF NOT EXISTS discogs (id INTEGER PRIMARY KEY, artist TEXT, title TEXT, year INTEGER, image TEXT)")
 
-def insert_fake_data(connection):
-    """Insert fake data into the tables."""
-    execute_sql(connection, "INSERT INTO ranking (id, name, description, url, style, year, date, language, created_date, active) VALUES ('test', 'Test', 'Test', 'http://test.com', 'Test', 2021, '2021-01-01', 'fr', '2021-01-01 00:00:00', 1)")
-    execute_sql(connection, "INSERT INTO album (id_ranking, rank, artist, title, year, id_discogs) VALUES ('test', 1, 'Test', 'Test', 2021, 1)")
-    execute_sql(connection, "INSERT INTO discogs (id, artist, title, year, image) VALUES (1, 'Test', 'Test', 2021, 'http://test.com')")
-
+def populate_database(connection):
+    # Insert fake ranking
+    execute_sql(connection, "INSERT INTO ranking (id, name, description, url, style, year, date, language, created_date, active) VALUES ('628cf492-5001-4e6c-9c3a-4db5d531f33a','Mes Tops', 'Mes 10 meilleurs albums de tous les temps', '', '',0,'2021-01-01','fr','2024-03-14', 1)")
+    # Insert fake album without discogs id
+    execute_sql(connection, "INSERT INTO album (id_ranking, rank, artist, title, year, id_discogs, reference) VALUES ( '628cf492-5001-4e6c-9c3a-4db5d531f33a', 1, 'The Cure', 'Disintegration', 1989, NULL, NULL)")    
+    execute_sql(connection, "INSERT INTO album (id_ranking, rank, artist, title, year, id_discogs, reference) VALUES ( '628cf492-5001-4e6c-9c3a-4db5d531f33a', 2, 'Joy Division', 'Closer', 1980, NULL, NULL)")    
+    execute_sql(connection, "INSERT INTO album (id_ranking, rank, artist, title, year, id_discogs, reference) VALUES ( '628cf492-5001-4e6c-9c3a-4db5d531f33a', 3, 'Nirvana', 'Nevermind', 0, NULL, NULL)")    
+        
 def create_database(database_path):
     # Create data directory if it does not exist
     data_directory = os.path.dirname(database_path)
@@ -37,11 +39,23 @@ def create_database(database_path):
     create_table_ranking(connection)
     create_table_album(connection)
     create_table_discogs(connection)
-    insert_fake_data(connection)
     connection.commit()
     connection.close()
     
+def create_test_database(database_path):
+    if os.path.exists(database_path):
+        drop_database(database_path)
+    create_database(database_path)
+    connection = sqlite3.connect(database_path)
+    populate_database(connection)
+    connection.close()
+    return 0
+       
 def drop_database(database_path):
+    # Check if the database exists
+    if not os.path.exists(database_path):
+        return -1
+
     # Drop the SQLite database
     os.remove(database_path)
     return 0
